requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cloudinary==1.33.0
Django==3.2
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==1.26.16
